requirements.txt
csv
os
datetime
collections
pandas
numpy
sklearn.linear_modeal
sklearn.model_sleection
sklearn.metrics